determine base types and interfaces
