+++ ts/.eslintrc
@@ -29,5 +29,8 @@
     "no-unused-expressions": "off",
     "class-methods-use-this": "off",
     "max-classes-per-file": "off"
-  }
+  },
+  "ignorePatterns": [
+    "generated.ts"
+  ]
 }
