+++ ts/package.json
@@ -28,24 +28,24 @@
     "start": "npm run test:watch"
   },
   "devDependencies": {
-    "@types/jest": "^29.1.1",
-    "@types/node": "^18.8.2",
-    "@typescript-eslint/eslint-plugin": "^5.39.0",
-    "@typescript-eslint/parser": "^5.39.0",
-    "concurrently": "^7.4.0",
-    "esbuild": "^0.15.10",
-    "eslint": "^8.24.0",
+    "@types/jest": "^29.2.4",
+    "@types/node": "^18.11.17",
+    "@typescript-eslint/eslint-plugin": "^5.46.1",
+    "@typescript-eslint/parser": "^5.46.1",
+    "concurrently": "^7.6.0",
+    "esbuild": "^0.16.9",
+    "eslint": "^8.30.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "^27.1.1",
+    "eslint-plugin-jest": "^27.1.7",
     "eslint-plugin-prettier": "^4.2.1",
-    "jest": "^29.1.2",
-    "prettier": "^2.7.1",
+    "jest": "^29.3.1",
+    "prettier": "^2.8.1",
     "shx": "^0.3.4",
     "ts-jest": "^29.0.3",
     "ts-node": "^10.9.1",
-    "tslib": "^2.4.0",
+    "tslib": "^2.4.1",
     "typescript": "^4.8.4"
   }
 }
