+++ ts/packages/ts/package.json
@@ -16,12 +16,10 @@
     "configu-typescript"
   ],
   "scripts": {
-    "prepare": "npm run build:types",
-    "build:types": "OUT_LANG=typescript sh ../../../jtd/generate.sh",
     "build": "tsc --build --verbose"
   },
   "dependencies": {
-    "ajv": "^8.11.0",
+    "ajv": "^8.11.2",
     "axios": "^0.27.2",
     "lodash": "^4.17.21",
     "mustache": "^4.2.0",
@@ -29,8 +27,8 @@
     "validator": "^13.7.0"
   },
   "devDependencies": {
-    "@types/lodash": "^4.14.186",
-    "@types/mustache": "^4.2.1",
-    "@types/validator": "^13.7.7"
+    "@types/lodash": "^4.14.191",
+    "@types/mustache": "^4.2.2",
+    "@types/validator": "^13.7.10"
   }
 }
