--- ts/packages/ts/src/types/.eslintrc
@@ -1,10 +1,0 @@
-{
-  "extends": [
-    "../../../../.eslintrc"
-  ],
-  "rules": {
-    "no-shadow": "off",
-    "no-use-before-define": "off",
-    "prettier/prettier": "off"
-  }
-}
