+++ types/ConfigSet.ts
@@ -1,0 +1,8 @@
+/**
+ * An interface of a path in an hierarchy, aka ConfigSet
+ * that contains Config <value> permutation
+ */
+export interface ConfigSet {
+  path: string;
+  hierarchy: string[];
+}
