--- types/SchemaContents.ts
@@ -1,5 +1,0 @@
-import { ConfigSchema } from "./ConfigSchema";
-
-export interface SchemaContentsValue extends ConfigSchema {};
-export type SchemaContents = { [key: string]: SchemaContentsValue };
-
